Feature/reports (#19)
* Modified public profile to make it more responsive.
Added constrain to report table to make accused and accuser unable to be the same person.
Removed transaction_id from the report table because we had agreed users would be able to report eachother without having been in a transaction together.
Created page to make a report.
Made link in public profile for "Report" send to the new report page.
Created individual report page showing details including messages in the report.

* Added send message form in report page

* Made report message update on UI after submitting without refreshing the page

* Made the public profile page more responsive.
Added a confirmation dialogue before deleting review on public profile as asked by the client.
Removed display of number of suspensions on public profile as asked by the client.